LayoutTests/crypto/webkitSubtle/aes-cbc-generate-key-expected.txt

CONSOLE MESSAGE: line 616: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test generating an AES key using AES-CBC algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.webkitSubtle.generateKey("aes-cbc", extractable, ["encrypt", "decrypt"]) threw exception TypeError: Type error.
PASS crypto.webkitSubtle.generateKey({name: "aes-cbc"}, extractable, ["encrypt", "decrypt"]) threw exception TypeError: Value NaN is outside the range [0, 65535].
PASS crypto.webkitSubtle.generateKey({name: "aes-cbc", length: undefined}, extractable, ["encrypt", "decrypt"]) threw exception TypeError: Value NaN is outside the range [0, 65535].
PASS crypto.webkitSubtle.generateKey({name: "aes-cbc", length: {}}, extractable, ["encrypt", "decrypt"]) threw exception TypeError: Value NaN is outside the range [0, 65535].
Generating a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ['decrypt', 'encrypt']
PASS successfullyParsed is true

TEST COMPLETE
